Containers: rename to .wrap.js
