+++ generators/react/templates/scripts/generators/component/index.js.hbs
@@ -3,5 +3,5 @@
 export { default } from './{{ properCase name }}.component';
 {{/unless}}
 {{#if wantConnect}}
-export { default } from './{{ properCase name }}.container';
+export { default } from './{{ properCase name }}.wrap';
 {{/if}}
