+++ generators/react/templates/scripts/generators/component/wrap.js.hbs
