+++ generators/react/templates/src/components/Root/Root.wrap.js
