+++ generators/react/templates/src/pages/Avatar/Avatar.wrap.js
