+++ generators/react/templates/src/pages/Avatar/index.js
@@ -1,2 +1,2 @@
 // @flow
-export { default } from './Avatar.container';
+export { default } from './Avatar.wrap';
