+++ generators/react/templates/src/pages/Home/Home.wrap.js
